+++ kripodb/canned.py
@@ -24,7 +24,8 @@ from requests import HTTPError
 
 from .db import FragmentsDb
 from .pairs import similar, open_similarity_matrix
-from .webservice.client import WebserviceClient, IncompleteFragments
+from .pharmacophores import PharmacophoresDb, as_phar
+from .webservice.client import WebserviceClient, IncompleteFragments, IncompletePharmacophores
 
 
 class IncompleteHits(Exception):
@@ -62,7 +63,8 @@ def similarities(queries, similarity_matrix_filename_or_url, cutoff, limit=1000)
         11
 
         Retrieved from web service instead of local similarity matrix file.
-        Make sure the web service is running, for example by `kripodb serve data/similarities.h5 data/fragments.sqlite`.
+        Make sure the web service is running,
+        for example by `kripodb serve data/similarities.h5 data/fragments.sqlite data/pharmacophores.h5`.
 
         >>> hits = similarities(queries, 'http://localhost:8084/kripo', 0.55)
         >>> len(hits)
@@ -134,7 +136,8 @@ def fragments_by_pdb_codes(pdb_codes, fragments_db_filename_or_url, prefix=''):
         3
 
         Retrieved from web service instead of local fragments db file.
-        Make sure the web service is running, for example by `kripodb serve data/similarities.h5 data/fragments.sqlite`.
+        Make sure the web service is running,
+        for example by `kripodb serve data/similarities.h5 data/fragments.sqlite data/pharmacophores.h5`.
 
         >>> fragments = fragments_by_pdb_codes(pdb_codes, 'http://localhost:8084/kripo')
         >>> len(fragments)
@@ -192,7 +195,8 @@ def fragments_by_id(fragment_ids, fragments_db_filename_or_url, prefix=''):
         1
 
         Retrieved from web service instead of local fragments db file.
-        Make sure the web service is running, for example by `kripodb serve data/similarities.h5 data/fragments.sqlite`.
+        Make sure the web service is running,
+        for example by `kripodb serve data/similarities.h5 data/fragments.sqlite data/pharmacophores.h5`.
 
         >>> fragments = fragments_by_id(fragment_ids,, 'http://localhost:8084/kripo')
         >>> len(fragments)
@@ -229,3 +233,56 @@ def fragments_by_id(fragment_ids, fragments_db_filename_or_url, prefix=''):
     df = pd.DataFrame(fragments)
     df.rename(columns=lambda x: prefix + x, inplace=True)
     return df
+
+
+def pharmacophores_by_id(fragment_ids, pharmacophores_db_filename_or_url):
+    """Fetch pharmacophore points by fragment identifiers
+
+    Args:
+        fragment_ids (pd.Series): List of fragment identifiers
+        pharmacophores_db_filename_or_url: Filename of pharmacophores db or base url of kripodb webservice
+
+    Returns:
+        pandas.Series: Pandas series with pharmacophores as string in phar format.
+                       Fragment without pharmacophore will return None
+
+    Examples:
+        Fragments similar to '3j7u_NDP_frag24' fragment.
+
+        >>> from kripodb.canned import pharmacophores_by_id
+        >>> fragment_ids = pd.Series(['2n2k_MTN_frag1'])
+        >>> pharmacophores = pharmacophores_by_id(fragment_ids, 'data/pharmacophores.h5')
+        >>> len(pharmacophores)
+        1
+
+        Retrieved from web service instead of local pharmacophores db file.
+        Make sure the web service is running,
+        for example by `kripodb serve data/similarities.h5 data/fragments.sqlite data/pharmacophores.h5`.
+
+        >>> pharmacophores = pharmacophores_by_id(fragment_ids,, 'http://localhost:8084/kripo')
+        >>> len(pharmacophores)
+        1
+    """
+    if pharmacophores_db_filename_or_url.startswith('http'):
+        client = WebserviceClient(pharmacophores_db_filename_or_url)
+        try:
+            pphors = client.pharmacophores(fragment_ids)
+        except IncompletePharmacophores as e:
+            s = pd.Series(e.pharmacophores, dtype=str)
+            raise IncompletePharmacophores(e.absent_identifiers, s)
+    else:
+        with PharmacophoresDb(pharmacophores_db_filename_or_url) as pharmacophoresdb:
+            pphors = []
+            absent_identifiers = []
+            for frag_id in fragment_ids:
+                try:
+                    phar = as_phar(frag_id, pharmacophoresdb[frag_id])
+                    pphors.append(phar)
+                except KeyError:
+                    pphors.append(None)
+                    absent_identifiers.append(frag_id)
+            if absent_identifiers:
+                s = pd.Series(pphors, dtype=str)
+                raise IncompletePharmacophores(absent_identifiers, s)
+    s = pd.Series(pphors, dtype=str)
+    return s